+++ .idea/misc.xml
@@ -5,7 +5,7 @@
       <map>
         <entry key="..\:/Users/irvin/AndroidStudioProjects/Pokedex/app/src/main/res/drawable-v24/card_background.xml" value="0.1745" />
         <entry key="..\:/Users/irvin/AndroidStudioProjects/Pokedex/app/src/main/res/layout/activity_main.xml" value="0.2421875" />
-        <entry key="..\:/Users/irvin/AndroidStudioProjects/Pokedex/app/src/main/res/layout/activity_pokemon_detail.xml" value="0.264" />
+        <entry key="..\:/Users/irvin/AndroidStudioProjects/Pokedex/app/src/main/res/layout/activity_pokemon_detail.xml" value="0.2" />
         <entry key="..\:/Users/irvin/AndroidStudioProjects/Pokedex/app/src/main/res/layout/fragment_about.xml" value="0.2421875" />
         <entry key="..\:/Users/irvin/AndroidStudioProjects/Pokedex/app/src/main/res/layout/fragment_evolution.xml" value="0.2421875" />
         <entry key="..\:/Users/irvin/AndroidStudioProjects/Pokedex/app/src/main/res/layout/fragment_stat.xml" value="0.2421875" />
