--- app/src/main/java/com/example/pokedex/ui/pokemonDetail/AboutFragment.kt
@@ -1,28 +1,0 @@
-package com.example.pokedex.ui.pokemonDetail
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.pokedex.R
-import com.example.pokedex.databinding.FragmentAboutBinding
-
-class AboutFragment : Fragment() {
-    private var _binding: FragmentAboutBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentAboutBinding.inflate(inflater, container, false)
-        return binding.root
-
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-
-}
