+++ app/src/main/java/com/example/pokedex/ui/pokemonDetail/adapter/ViewPageAdapter.kt
@@ -4,11 +4,13 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.example.pokedex.ui.pokemonDetail.AboutFragment
-import com.example.pokedex.ui.pokemonDetail.EvolutionFragment
-import com.example.pokedex.ui.pokemonDetail.StatFragment
+import com.example.pokedex.ui.pokemonDetail.fragments.AboutFragment.AboutFragment
+import com.example.pokedex.ui.pokemonDetail.fragments.EvolutionFragment.EvolutionFragment
+import com.example.pokedex.ui.pokemonDetail.fragments.StatFragment.StatFragment
 
-internal class ViewPageAdapter(fragmentManager: FragmentManager, lifecycle: Lifecycle): FragmentStateAdapter(fragmentManager, lifecycle){
+internal class ViewPageAdapter(fragmentManager: FragmentManager, lifecycle: Lifecycle, pokemonName: String): FragmentStateAdapter(fragmentManager, lifecycle){
+    private val _pokemonName = pokemonName
+
     override fun getItemCount(): Int {
         return 3
     }
@@ -16,7 +18,7 @@ internal class ViewPageAdapter(fragmentManager: FragmentManager, lifecycle: Life
     override fun createFragment(position: Int): Fragment {
        return when (position){
             0-> {
-                AboutFragment()
+                AboutFragment(_pokemonName)
             }
             1-> {
                 StatFragment()
