+++ app/src/main/java/com/example/pokedex/ui/pokemonDetail/fragments/EvolutionFragment/EvolutionFragment.kt
@@ -1,12 +1,10 @@
-package com.example.pokedex.ui.pokemonDetail
+package com.example.pokedex.ui.pokemonDetail.fragments.EvolutionFragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.pokedex.R
-import com.example.pokedex.databinding.FragmentAboutBinding
 import com.example.pokedex.databinding.FragmentEvolutionBinding
