+++ app/src/main/res/layout/fragment_evolution.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.pokemonDetail.EvolutionFragment">
+    tools:context=".ui.pokemonDetail.fragments.EvolutionFragment.EvolutionFragment">
 
     <TextView
         android:layout_width="match_parent"
